requirements.txt
boto3
xarray
netCDF4
scipy
cfgrib
pyyaml
numpy
pandas
eccodes
